Incorporate minor PR feedback re: tests for FingerprintClientHello work
